% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{CalculateSaturationFor10x}
\alias{CalculateSaturationFor10x}
\title{Calculate Saturation For 10x}
\usage{
CalculateSaturationFor10x(barcodeMatrix, molInfoFile, doPlot = TRUE)
}
\arguments{
\item{barcodeMatrix}{The matrix holding HTO count data. The columns should match cells}

\item{molInfoFile}{The 10x molecule_info.h5 file}

\item{doPlot}{If true, plots summarizing saturation will be generated.}

\item{cellbarcodePrefix}{An optional string appended to the barcodes parsed from the molecule_info.h5 file. This is necessary if the seurat object has a prefix applied to cell barcodes. This value is directly concatenated and must include any delimiter. Note: if this is absent, but the seuratObj has the columns DatasetId or BarcodePrefix, the latter will be used.}
}
\value{
A dataframe with the cellbarcode and per-cell saturation
}
\description{
Calculate per-cell saturation for 10x data
}
