% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{CallAndGenerateReport}
\alias{CallAndGenerateReport}
\title{CallAndGenerateReport}
\usage{
CallAndGenerateReport(
  rawCountData,
  reportFile,
  callFile,
  barcodeWhitelist = NULL,
  cellbarcodeWhitelist = "matrix"
)
}
\arguments{
\item{rawCountData}{The input barcode file or umi_count folder}

\item{reportFile}{The file to which the HTML report will be written}

\item{callFile}{The file to which the table of calls will be written}

\item{barcodeWhitelist}{A vector of barcode names to retain.}

\item{cellbarcodeWhitelist}{Either a vector of expected barcodes (such as all cells with passing gene expression data), or the string 'matrix'. If the latter is provided, the set of cellbarcodes present in the original unfiltered count matrix will be stored and used for reporting. This allows the report to count cells that were filtered due to low counts separately from negative/non-callable cells.}

\item{minCountPerCell}{Cells (columns) will be dropped if their total count is less than this value.}
}
\description{
Runs the default processing pipeline
}
