% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ConfigureGMMDemux}
\alias{ConfigureGMMDemux}
\title{Configure GMMDemux}
\usage{
ConfigureGMMDemux()
}
\description{
GMMDemux is a python package used for calling cell hashing. If not already present, this function will automatically create a virtualenv and install GMM_Demux into it. Note: you need to call reticulate::py_config() prior to using this.
}
