% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{DoMULTIseqDemux}
\alias{DoMULTIseqDemux}
\title{Perform HTO classification using Seurat's implementation of MULTI-seq classification}
\usage{
DoMULTIseqDemux(
  seuratObj,
  assay = "HTO",
  autoThresh = TRUE,
  quantile = NULL,
  maxiter = 20,
  qrange = seq(from = 0.2, to = 0.95, by = 0.05)
)
}
\arguments{
\item{seuratObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
Perform HTO classification using Seurat's implementation of MULTI-seq classification
}
