% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFF_Demux.R
\name{.EstimateMultipletRate}
\alias{.EstimateMultipletRate}
\title{Estimate Multiplet Rate}
\usage{
.EstimateMultipletRate(barcodeMatrix, num10xRuns = 1)
}
\arguments{
\item{barcodeMatrix}{a barcodes-by-cells matrix of hashing counts}

\item{num10xRuns}{The number of lanes the cells are loaded into (expected to be 1)}
}
\description{
Estimates multiplet rate by using barcode matrix dimensions and formulae from the Saijita Lab's calculator (https://satijalab.org/costpercell/)
}
