% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{DownloadAndAppendCellHashing}
\alias{DownloadAndAppendCellHashing}
\title{DownloadAndAppendCellHashing}
\usage{
DownloadAndAppendCellHashing(seuratObject, outPath = ".")
}
\arguments{
\item{seuratObject, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
Downloads matching Cell Hashing data using barcodePrefix on the seurat object and appends it to metadata
}
