% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{EstimateMultipletRate}
\alias{EstimateMultipletRate}
\title{Estimate Multiplet Rate for 10x Genomics Data}
\usage{
EstimateMultipletRate(numCellsRecovered, num10xRuns = 1, chemistry = "10xV3")
}
\arguments{
\item{numCellsRecovered}{The number of cells recovered (number of columns in the count matrix)}

\item{num10xRuns}{The number of lanes the cells are loaded into (expected to be 1)}

\item{chemistry}{The version of 10x chemistry reagent kit used}
}
\description{
Estimates multiplet rate by using barcode matrix dimensions and formulae from the Saijita Lab's calculator (https://satijalab.org/costpercell/)
}
