% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateCellHashCallsMultiSeq}
\alias{GenerateCellHashCallsMultiSeq}
\title{A Title}
\usage{
GenerateCellHashCallsMultiSeq(barcodeMatrix, method = "seurat")
}
\value{
A modified Seurat object.
}
\description{
A description
}
