% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateCellHashCallsSeurat}
\alias{GenerateCellHashCallsSeurat}
\title{GenerateCellHashCallsSeurat}
\usage{
GenerateCellHashCallsSeurat(
  barcodeMatrix,
  positive.quantile = 0.99,
  attemptRecovery = F,
  minCellsForRecovery = 500
)
}
\value{
A data frame of results
}
\description{
Generates final cell hashing calls using Seurat3 HTODemux
}
