% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateCellHashingCalls}
\alias{GenerateCellHashingCalls}
\title{GenerateCellHashingCalls}
\usage{
GenerateCellHashingCalls(
  barcodeMatrix,
  htodemux.positive.quantile = 0.95,
  methods = c("htodemux", "multiseq")
)
}
\value{
A data table of results.
}
\description{
A description
}
