% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateSummaryForExpectedBarcodes}
\alias{GenerateSummaryForExpectedBarcodes}
\title{GenerateSummaryForExpectedBarcodes}
\usage{
GenerateSummaryForExpectedBarcodes(
  dt,
  whitelistFile,
  outputFile,
  barcodeMatrix
)
}
\description{
A description
}
