% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GetExampleMarkdown}
\alias{GetExampleMarkdown}
\title{GetExampleMarkdown}
\usage{
GetExampleMarkdown(dest)
}
\arguments{
\item{dest}{The destination filepath, where the file will be saved}
}
\description{
Save a template R markdown file, showing usage of this package
}
