% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_HTO_Demux.R
\name{HTODemux2}
\alias{HTODemux2}
\title{HTODemux2}
\usage{
HTODemux2(
  object,
  assay = "HTO",
  positive.quantile = 0.98,
  nstarts = 100,
  kfunc = "clara",
  nsamples = 100,
  verbose = TRUE,
  plotDist = FALSE
)
}
\arguments{
\item{object, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
\keyword{SerIII_template}
