% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R, R/Visualization.R
\name{HtoSummary}
\alias{HtoSummary}
\title{A Title}
\usage{
HtoSummary(
  seuratObj,
  htoClassificationField,
  globalClassificationField,
  label,
  doHeatmap = T,
  doTSNE = T,
  assay = "HTO"
)

HtoSummary(
  seuratObj,
  htoClassificationField,
  globalClassificationField,
  label,
  doHeatmap = T,
  doTSNE = T,
  assay = "HTO"
)
}
\arguments{
\item{seuratObj, }{A Seurat object.}
}
\value{
A modified Seurat object.

A modified Seurat object.
}
\description{
A description

A description
}
