% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{LoadCiteSeqCountData}
\alias{LoadCiteSeqCountData}
\title{Load and Plot CiteSeq Count Data}
\usage{
LoadCiteSeqCountData(rawCountData = NA)
}
\arguments{
\item{rawCountData, }{The input barcode file or umi_count folder}
}
\description{
Load and create QC plots for CITE-seq data
}
