% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{PlotLibrarySaturationByMarker}
\alias{PlotLibrarySaturationByMarker}
\title{Plot Library Saturation By Marker}
\usage{
PlotLibrarySaturationByMarker(citeseqCountDir)
}
\arguments{
\item{citeseqCountDir}{The root of the Cite-seq-Count output folder, which should contain umi_count and read_count folders.}
}
\description{
Create a plot of the library saturation per cell, separated by marker
}
