% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{PlotNormalizationQC}
\alias{PlotNormalizationQC}
\title{Plot Normalization QC}
\usage{
PlotNormalizationQC(
  barcodeData,
  methods = c("bimodalQuantile", "Quantile", "log2Center", "CLR")
)
}
\arguments{
\item{barcodeData}{The count matrix}

\item{methods}{The normalizations to use. Allowable options are: bimodalQuantile, Quantile, log2Center, CLR and Raw}
}
\description{
Generates QC plots related to normalization
}
