% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{PrintFinalSummary}
\alias{PrintFinalSummary}
\title{PrintFinalSummary}
\usage{
PrintFinalSummary(df, barcodeMatrix)
}
\arguments{
\item{The}{barcode counts matrix}
}
\value{
A modified Seurat object.
}
\description{
PrintFinalSummary
}
