% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{ProcessCountMatrix}
\alias{ProcessCountMatrix}
\title{Process CiteSeq Count Matrix}
\usage{
ProcessCountMatrix(
  rawCountData = NA,
  minCountPerCell = 5,
  barcodeWhitelist = NULL,
  barcodeBlacklist = c("no_match", "total_reads", "unmapped"),
  cellbarcodeWhitelist = NULL,
  doPlot = TRUE,
  simplifyBarcodeNames = TRUE,
  saveOriginalCellBarcodeFile = NULL,
  metricsFile = NULL
)
}
\arguments{
\item{rawCountData, }{The input barcode file or umi_count folder}

\item{minCountPerCell}{Cells (columns) will be dropped if their total count is less than this value.}

\item{barcodeWhitelist}{A vector of barcode names to retain.}

\item{barcodeBlacklist}{A vector of barcodes names to discard.}

\item{cellbarcodeWhitelist}{If provided, the raw count matrix will be subset to include only these cells. This allows one to use the cellranger unfiltered matrix as an input, but filter based on target cells, such as those with GEX data.}

\item{doPlot}{If true, QC plots will be generated}

\item{simplifyBarcodeNames}{If true, the sequence tag portion will be removed from the barcode names (i.e. HTO-1-ATGTGTGA -> HTO-1)}

\item{saveOriginalCellBarcodeFile}{An optional file path, where the set of original cell barcodes, prior to filtering, will be written. The primary use-case is if the count matrix was generated using a cell whitelist (like cells with passing gene expression). Preserving this list allows downstream reporting.}

\item{metricsFile}{If provided, summary metrics will be written to this file.}
}
\value{
The updated count matrix
}
\description{
The primary entrypoint for parsing and QC of the cell hashing count matrix.
}
