% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{ProcessEnsemblHtoCalls}
\alias{ProcessEnsemblHtoCalls}
\title{ProcessEnsemblHtoCalls}
\usage{
ProcessEnsemblHtoCalls(
  mc,
  sc,
  barcodeMatrix,
  outFile = "combinedHtoCalls.txt",
  allCallsOutFile = NA
)
}
\arguments{
\item{mc}{Multiseq calls dataframe}

\item{sc}{Seurat calls dataframe}

\item{barcodeMatrix}{The barcode count matrix}

\item{outFile}{The output TSV file}

\item{allCallsOutFile}{If provided, a more detailed output will be written here}
}
\value{
A modified Seurat object.
}
\description{
A description
}
