% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{SetSeed}
\alias{SetSeed}
\title{Set random seed}
\usage{
SetSeed(seed)
}
\arguments{
\item{seed}{The random seed}
}
\description{
Sets the seed used for R‘s random number generator, which should be used in all internal functions
}
