% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{SummarizeCellsByClassification}
\alias{SummarizeCellsByClassification}
\title{SummarizeCellsByClassification}
\usage{
SummarizeCellsByClassification(calls, barcodeMatrix)
}
\arguments{
\item{barcodeMatrix}{The filtered matrix of hashing count data}
}
\description{
Create summary plots to contrast cells based on call-status. This is designed to help inform why specific cells were not called.
}
