% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{CalculateSaturationFor10x}
\alias{CalculateSaturationFor10x}
\title{Calculate Saturation For 10x}
\usage{
CalculateSaturationFor10x(barcodeMatrix, molInfoFile, doPlot = TRUE)
}
\arguments{
\item{barcodeMatrix}{The matrix holding HTO count data. The columns should match cells}

\item{molInfoFile}{The 10x molecule_info.h5 file}

\item{doPlot}{If true, plots summarizing saturation will be generated.}
}
\value{
A dataframe with the cellbarcode and per-cell saturation
}
\description{
Calculate per-cell saturation for 10x data
}
