% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{CallAndGenerateReport}
\alias{CallAndGenerateReport}
\title{Call Cell Hashing And Generate Report}
\usage{
CallAndGenerateReport(
  rawCountData,
  reportFile,
  callFile,
  h5File = NULL,
  barcodeWhitelist = NULL,
  cellbarcodeWhitelist = "inputMatrix",
  methods = c("multiseq", "htodemux"),
  methodsForConsensus = NULL,
  minCountPerCell = 5,
  title = NULL,
  metricsFile = NULL,
  rawCountsExport = NULL,
  skipNormalizationQc = FALSE,
  keepMarkdown = FALSE,
  molInfoFile = NULL,
  majorityConsensusThreshold = NULL,
  callerDisagreementThreshold = NULL
)
}
\arguments{
\item{rawCountData}{The input barcode file or umi_count folder}

\item{reportFile}{The file to which the HTML report will be written}

\item{callFile}{The file to which the table of calls will be written}

\item{h5File}{demuxEM requires the 10x h5 gene expression count file. This is only required when demuxEM is used.}

\item{barcodeWhitelist}{A vector of barcode names to retain.}

\item{cellbarcodeWhitelist}{Either a vector of expected barcodes (such as all cells with passing gene expression data), a file with one cellbarcode per line, or the string 'inputMatrix'. If the latter is provided, the set of cellbarcodes present in the original unfiltered count matrix will be stored and used for reporting. This allows the report to count cells that were filtered due to low counts separately from negative/non-callable cells.}

\item{methods}{The set of methods to use for calling. See GenerateCellHashingCalls for options.}

\item{methodsForConsensus}{By default, a consensus call will be generated using all methods; however, if this parameter is provided, all algorithms specified by methods will be run, but only the list here will be used for the final consensus call. This allows one to see the results of a given caller without using it for the final calls.}

\item{minCountPerCell}{Cells (columns) will be dropped if their total count is less than this value.}

\item{title}{A title for the HTML report}

\item{metricsFile}{If provided, summary metrics will be written to this file.}

\item{rawCountsExport}{If provided, the raw count matrix, after processing, will be written as an RDS object to this file. This can be useful for debugging.}

\item{skipNormalizationQc}{If true, the normalization/QC plots will be skipped. These can be time consuming on large input data.}

\item{keepMarkdown}{If true, the markdown file will be saved, in addition to the HTML file}

\item{molInfoFile}{An optional path to the 10x molecule_info.h5.}

\item{majorityConsensusThreshold}{This applies to calculating a consensus call when multiple algorithms are used. If NULL, then all non-negative calls must agree or that cell is marked discordant. If non-NULL, then the number of algorithms returning the top call is divided by the total number of non-negative calls. If this ratio is above the majorityConsensusThreshold, that value is selected. For example, when majorityConsensusThreshold=0.6 and the calls are: HTO-1,HTO-1,Negative,HTO-2, then 2/3 calls are for HTO-1, giving 0.66. This is greater than the majorityConsensusThreshold of 0.6, so HTO-1 is returned. This can be useful for situations where most algorithms agree, but a single caller fails.}

\item{callerDisagreementThreshold}{If provided, the agreement rate will be calculated between each caller and the simple majority call, ignoring discordant and no-call cells. If any caller has an disagreement rate above this threshold, it will be dropped and the consensus call re-calculated. The general idea is to drop a caller that is systematically discordant.}

\item{doTSNE}{If true, tSNE will be run on results as part of QC. This can be memory intensive and is not strictly needed, so it can be skipped if desired.}
}
\description{
Runs the default processing pipeline
}
