% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{CallAndGenerateReport}
\alias{CallAndGenerateReport}
\title{CallAndGenerateReport}
\usage{
CallAndGenerateReport(
  rawCountData,
  reportFile,
  callFile,
  barcodeWhitelist = NULL,
  cellbarcodeWhitelist = "inputMatrix",
  methods = c("multiseq", "htodemux"),
  citeSeqCountDir = NULL,
  minCountPerCell = 5,
  title = NULL
)
}
\arguments{
\item{rawCountData}{The input barcode file or umi_count folder}

\item{reportFile}{The file to which the HTML report will be written}

\item{callFile}{The file to which the table of calls will be written}

\item{barcodeWhitelist}{A vector of barcode names to retain.}

\item{cellbarcodeWhitelist}{Either a vector of expected barcodes (such as all cells with passing gene expression data), or the string 'inputMatrix'. If the latter is provided, the set of cellbarcodes present in the original unfiltered count matrix will be stored and used for reporting. This allows the report to count cells that were filtered due to low counts separately from negative/non-callable cells.}

\item{methods}{The set of methods to use for calling. See GenerateCellHashingCalls for options.}

\item{citeSeqCountDir}{This is the root folder of the Cite-seq-Count output, containing umi_count and read_count folders. If provided, this will be used to generate a library saturation plot}

\item{minCountPerCell}{Cells (columns) will be dropped if their total count is less than this value.}

\item{title}{A title for the HTML report}
}
\description{
Runs the default processing pipeline
}
