% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{DebugDemux}
\alias{DebugDemux}
\title{A Title}
\usage{
DebugDemux(seuratObj, assay = "HTO", reportKmeans = FALSE)
}
\arguments{
\item{seuratObj, }{A Seurat object.}
}
\value{
A modified Seurat object.
}
\description{
A description
}
