% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_HTO_Demux.R
\name{GenerateCellHashCallsSeurat}
\alias{GenerateCellHashCallsSeurat}
\title{GenerateCellHashCallsSeurat}
\usage{
GenerateCellHashCallsSeurat(barcodeMatrix, positive.quantile = 0.95)
}
\value{
A data frame of results
}
\description{
Generates cell hashing calls using Seurat's HTODemux
}
