% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateCellHashingCalls}
\alias{GenerateCellHashingCalls}
\title{Generate Cell Hashing Calls}
\usage{
GenerateCellHashingCalls(
  barcodeMatrix,
  methods = c("bff_cluster", "multiseq", "dropletutils"),
  cellbarcodeWhitelist = NULL,
  metricsFile = NULL,
  doTSNE = TRUE,
  doHeatmap = TRUE,
  ...
)
}
\arguments{
\item{barcodeMatrix}{The filtered matrix of hashing count data}

\item{methods}{A vector of one or more calling methods to use. Currently supported are: htodemux, multiseq, dropletutils, gmm_demux, demuxem, bff_raw, and bff_cluster}

\item{cellbarcodeWhitelist}{A vector of expected cell barcodes. This allows reporting on the total set of expected barcodes, not just those in the filtered count matrix.}

\item{metricsFile}{If provided, summary metrics will be written to this file.}

\item{doTSNE}{If true, tSNE will be run on the resulting hashing calls after each caller. This can be useful as a sanity check; however, adds time.}

\item{doHeatmap}{If true, Seurat::HTOHeatmap will be run on the results of each caller. Not supported by all callers.}

\item{\dots}{Caller-specific arguments can be passed by prefixing with the method name. For example, htodemux.positive.quantile = 0.95, will be passed to the htodemux positive.quantile argument).}
}
\value{
A data frame of results.
}
\description{
The primary methods to generating cell hashing calls from a filtered matrix of count data.
}
