% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateCellHashingCalls}
\alias{GenerateCellHashingCalls}
\title{GenerateCellHashingCalls}
\usage{
GenerateCellHashingCalls(
  barcodeMatrix,
  positive.quantile = 0.99,
  attemptRecovery = FALSE,
  useSeurat = TRUE,
  useMultiSeq = TRUE,
  outFile = "combinedHtoCalls.txt",
  allCallsOutFile = NA
)
}
\value{
A data table of results.
}
\description{
A description
}
