% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateCellHashingCalls}
\alias{GenerateCellHashingCalls}
\title{GenerateCellHashingCalls}
\usage{
GenerateCellHashingCalls(
  barcodeMatrix,
  methods = c("htodemux", "multiseq"),
  htodemux.positive.quantile = 0.95
)
}
\arguments{
\item{barcodeMatrix}{The filtered matrix of hashing count data}

\item{methods}{A vector of one or more calling methods to use. Currently supported are: htodemux and multiseq}
}
\value{
A data frame of results.
}
\description{
The primary methods to generating cell hashing calls from a filtered matrix of count data
}
