% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{GenerateCellHashingCalls}
\alias{GenerateCellHashingCalls}
\title{Generate Cell Hashing Calls}
\usage{
GenerateCellHashingCalls(
  barcodeMatrix,
  methods = c("multiseq", "dropletutils"),
  cellbarcodeWhitelist = NULL,
  metricsFile = NULL,
  ...
)
}
\arguments{
\item{barcodeMatrix}{The filtered matrix of hashing count data}

\item{methods}{A vector of one or more calling methods to use. Currently supported are: htodemux and multiseq}

\item{cellbarcodeWhitelist}{A vector of expected cell barcodes. This allows reporting on the total set of expected barcodes, not just those in the filtered count matrix.}

\item{metricsFile}{If provided, summary metrics will be written to this file.}
}
\value{
A data frame of results.
}
\description{
The primary methods to generating cell hashing calls from a filtered matrix of count data. Note: caller-specific arguments can be passed by prefixing with the method name (i.e. htodemux.positive.quantile = 0.95, will be passed to the htodemux positive.quantile argument)
}
