% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{PlotLibrarySaturation}
\alias{PlotLibrarySaturation}
\title{PlotLibrarySaturation}
\usage{
PlotLibrarySaturation(citeseqCountDir)
}
\arguments{
\item{citeseqCountDir, }{The root of the Cite-seq-Count output folder, which should contain umi_count and read_count folders.}
}
\value{
The overall saturation for this library
}
\description{
Create a plot of the library saturation per cell
}
