% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{PlotNormalizationQC}
\alias{PlotNormalizationQC}
\title{Plot Normalization QC}
\usage{
PlotNormalizationQC(barcodeData)
}
\arguments{
\item{barcodeData}{The count matrix}
}
\description{
Generates QC plots related to normalization
}
