% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{ProcessCountMatrix}
\alias{ProcessCountMatrix}
\title{ProcessCiteSeqCount}
\usage{
ProcessCountMatrix(
  rawCountData = NA,
  minCountPerCell = 5,
  barcodeWhitelist = NULL,
  barcodeBlacklist = c("no_match", "total_reads", "unmapped"),
  doPlot = T
)
}
\arguments{
\item{rawCountData, }{The input barcode file.}

\item{minCountPerCell}{Cells (columns) will be dropped if their total count is less than this value.}
}
\value{
The updated count matrix
}
\description{
The primary entrypoint for parsing and QC of the cell hashing count matrix.
}
