% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{ProcessEnsemblHtoCalls}
\alias{ProcessEnsemblHtoCalls}
\title{ProcessEnsemblHtoCalls}
\usage{
ProcessEnsemblHtoCalls(callList, barcodeMatrix, cellbarcodeWhitelist = NA)
}
\arguments{
\item{callList}{A list of dataframes, produced by callers}

\item{barcodeMatrix}{The barcode count matrix}
}
\description{
ProcessEnsemblHtoCalls
}
