% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellHashing.R
\name{ProcessEnsemblHtoCalls}
\alias{ProcessEnsemblHtoCalls}
\title{ProcessEnsemblHtoCalls}
\usage{
ProcessEnsemblHtoCalls(callList, cellbarcodeWhitelist = NULL)
}
\arguments{
\item{callList}{A list of data frames, produced by callers}

\item{cellbarcodeWhitelist}{A vector of expected cell barcodes. This allows reporting on the total set of expected barcodes, not just those in the filtered count matrix.}
}
\description{
ProcessEnsemblHtoCalls
}
